reorganised code base
