+++ Makefile
@@ -1,14 +1,22 @@
 
-GPPPARAMS	= -m32 -fno-use-cxa-atexit -nostdlib -fno-builtin -fno-rtti -fno-exceptions -fno-leading-underscore -Wno-write-strings
+GPPPARAMS	= -m32 -Iinclude -fno-use-cxa-atexit -nostdlib -fno-builtin -fno-rtti -fno-exceptions -fno-leading-underscore -Wno-write-strings
 ASPARAMS	= --32
 LDPARAMS	= -melf_i386
 
-objects = loader.o gdt.o port.o interruptstubs.o interrupts.o keyboard.o kernel.o
+objects = 	obj/loader.o \
+			obj/gdt.o \
+			obj/hardwarecommunication/port.o \
+			obj/hardwarecommunication/interruptstubs.o \
+			obj/hardwarecommunication/interrupts.o \
+			obj/drivers/keyboard.o \
+			obj/kernel.o
 
-%.o: %.cpp
+obj/%.o: src/%.cpp
+	mkdir -p $(@D)
 	g++ $(GPPPARAMS) -o $@ -c $<
 
-%.o: %.s
+obj/%.o: src/%.s
+	mkdir -p $(@D)
 	as $(ASPARAMS) -o $@ $<
 
 32bitkernel.bin: linker.ld $(objects)
@@ -40,4 +48,4 @@ run: 32bitkernel.iso
 
 .PHONY: clean
 clean:
-	rm -rf $(objects) 32bitkernel.bin 32bitkernel.iso
+	rm -rf obj 32bitkernel.bin 32bitkernel.iso
