+++ include/common/types.h
@@ -1,0 +1,24 @@
+
+#ifndef __TYPES_H
+#define __TYPES_H
+
+
+namespace thirty2bitOS
+{
+    namespace common
+    {
+        typedef char int8_t;
+        typedef unsigned char uint8_t;
+        
+        typedef short int16_t;
+        typedef unsigned short uint16_t;
+
+        typedef int int32_t;
+        typedef unsigned int uint32_t;
+
+        typedef long long int int64_t;
+        typedef unsigned long long uint64_t;
+    }
+}
+
+#endif
