--- keyboard.h
@@ -1,19 +1,0 @@
-#ifndef __KEYBOARD_H
-#define __KEYBOARD_H
-
-#include "types.h"
-#include "interrupts.h"
-#include "port.h"
-
-class KeyboardDriver : public InterruptHandler
-{
-    Port8Bit dataport;
-    Port8Bit commandport;
-
-    public:
-        KeyboardDriver(InterruptManager* manager);
-        ~KeyboardDriver();
-        virtual uint32_t handleInterrupt(uint32_t esp);
-};
-
-#endif
